add authentication to delete $ edit $ get message

+++ server/src/main/java/com/example/server/ServerController.java
@@ -9,15 +9,19 @@ import com.example.server.repository.ChatRepository;
 import com.example.server.repository.MessageRepository;
 import lombok.AllArgsConstructor;
 import com.example.server.repository.UserRepository;
+import org.apache.commons.lang.RandomStringUtils;
 import org.example.model.ChatModel;
 import org.example.model.GroupModel;
 import org.example.model.MessageModel;
 import org.example.model.PvModel;
+import org.example.model.UserModel;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -33,23 +37,25 @@ public class ServerController {
     private final UserRepository userRepository;
     private final MessageRepository messageRepository;
     private final ChatRepository chatRepository;
+    private static final int TOKEN_SIZE = 16;
 
     @PostMapping("/signUp")
     public ResponseEntity<String> signUp(@RequestBody String username) {
         if (userRepository.existsById(username)) {
             return new ResponseEntity<>(Commands.USERNAME_ALREADY_EXISTS, HttpStatus.BAD_REQUEST);
         }
-        userRepository.save(new User(username));
+        userRepository.save(new User(username, generateToken()));
         return new ResponseEntity<>(Commands.SUCCESSFULLY_SIGNED_UP, HttpStatus.OK);
     }
 
+    public static String generateToken() {
+        return RandomStringUtils.randomAlphabetic(TOKEN_SIZE);
+    }
+
     @GetMapping("signIn/{username}")
-    public ResponseEntity<String> signIn(@PathVariable String username) {
-        if (userRepository.existsById(username)) {
-            return new ResponseEntity<>(Commands.SUCCESSFULLY_SIGNED_IN, HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(Commands.USERNAME_DOESNT_EXIST, HttpStatus.BAD_REQUEST);
-        }
+    public ResponseEntity<UserModel> signIn(@PathVariable String username) {
+        Optional<User> user = userRepository.findById(username);
+        return user.map(value -> new ResponseEntity<>(value.createUserModel(), HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(null, HttpStatus.BAD_REQUEST));
     }
 
     @GetMapping("/chats/{username}")
@@ -137,20 +143,35 @@ public class ServerController {
     }
 
     @DeleteMapping("/messages/{messageId}")
-    public ResponseEntity<String> deleteMessage(@PathVariable Long messageId) {
+    public ResponseEntity<String> deleteMessage(@PathVariable Long messageId, @RequestHeader(HttpHeaders.AUTHORIZATION) String token) {
         Optional<Message> message = messageRepository.findById(messageId);
+        Optional<User> user = userRepository.findByToken(token);
+        if (user.isEmpty()) {
+            return new ResponseEntity<>(Commands.INVALID_TOKEN, HttpStatus.BAD_REQUEST);
+        }
         if (message.isEmpty()) {
             return new ResponseEntity<>(Commands.MESSAGE_DOESNT_EXIST, HttpStatus.BAD_REQUEST);
+        }
+        if (!message.get().getSender().equals(user.get().getUsername())) {
+            return new ResponseEntity<>(Commands.MESSAGE_DOESNT_YOUR_MESSAGE, HttpStatus.BAD_REQUEST);
         }
         messageRepository.deleteById(messageId);
         return new ResponseEntity<>(Commands.MESSAGE_DELETED, HttpStatus.OK);
     }
 
     @PutMapping("/messages/{messageId}")
-    public ResponseEntity<String> editMessage(@PathVariable Long messageId, @RequestBody MessageModel messageModel) {
+    public ResponseEntity<String> editMessage(@PathVariable Long messageId, @RequestBody MessageModel messageModel,
+                                              @RequestHeader(HttpHeaders.AUTHORIZATION) String token) {
+        Optional<User> user = userRepository.findByToken(token);
+        if (user.isEmpty()) {
+            return new ResponseEntity<>(Commands.INVALID_TOKEN, HttpStatus.BAD_REQUEST);
+        }
         Optional<Message> message = messageRepository.findById(messageId);
         if (message.isEmpty()) {
             return new ResponseEntity<>(Commands.MESSAGE_DOESNT_EXIST, HttpStatus.BAD_REQUEST);
+        }
+        if (!message.get().getSender().equals(user.get().getUsername())) {
+            return new ResponseEntity<>(Commands.MESSAGE_DOESNT_YOUR_MESSAGE, HttpStatus.BAD_REQUEST);
         }
         message.get().setText(messageModel.getText());
         messageRepository.save(message.get());
@@ -158,9 +179,13 @@ public class ServerController {
     }
 
     @GetMapping("/messages/{messageId}")
-    public ResponseEntity<Message> getMessage(@PathVariable Long messageId) {
+    public ResponseEntity<Message> getMessage(@PathVariable Long messageId, @RequestHeader(HttpHeaders.AUTHORIZATION) String token) {
+        Optional<User> user = userRepository.findByToken(token);
         Optional<Message> message = messageRepository.findById(messageId);
-        return message.map(value -> new ResponseEntity<>(value, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(null, HttpStatus.BAD_REQUEST));
+        if (user.isEmpty() || message.isEmpty() || !message.get().getSender().equals(user.get().getUsername())) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(message.get(), HttpStatus.OK);
     }
 
 }
